// src/app/models/cancer.ts
export interface Cancer {
    ID_CUENTA_CANCER?:string;
    CAMPO_1?: string;
    CAMPO_2?: string;
    CAMPO_3?: string;
    CAMPO_4?: string;
    CAMPO_5?: string;
    CAMPO_6?: string;
    CAMPO_7?: string;
    CAMPO_8?: string;
    CAMPO_9?: string;
    CAMPO_10?: string;
    CAMPO_11?: string;
    CAMPO_12?: string;
    CAMPO_13?: string;
    CAMPO_14?: string;
    CAMPO_15?: string;
    CAMPO_16?: string;
    CAMPO_17?: string;
    CAMPO_18?: string;
    CAMPO_19?: string;
    CAMPO_20?: string;
    CAMPO_21?: string;
    CAMPO_22?: string;
    CAMPO_23?: string;
    CAMPO_24?: string;
    CAMPO_25?: string;
    CAMPO_26?: string;
    CAMPO_27?: string;
    CAMPO_28?: string;
    CAMPO_29?: string;
    CAMPO_30?: string;
    CAMPO_31?: string;
    CAMPO_32?: string;
    CAMPO_33?: string;
    CAMPO_34?: string;
    CAMPO_35?: string;
    CAMPO_36?: string;
    CAMPO_37?: string;
    CAMPO_38?: string;
    CAMPO_39?: string;
    CAMPO_40?: string;
    CAMPO_41?: string;
    CAMPO_42?: string;
    CAMPO_43?: string;
    CAMPO_44?: string;
    CAMPO_45?: string;
    CAMPO_46?: string;
    CAMPO_46_1?: string;
    CAMPO_46_2?: string;
    CAMPO_46_3?: string;
    CAMPO_46_4?: string;
    CAMPO_46_5?: string;
    CAMPO_46_6?: string;
    CAMPO_46_7?: string;
    CAMPO_46_8?: string;
    CAMPO_47?: string;
    CAMPO_48?: string;
    CAMPO_49?: string;
    CAMPO_50?: string;
    CAMPO_51?: string;
    CAMPO_52?: string;
    CAMPO_53?: string;
    CAMPO_53_1?: string;
    CAMPO_53_2?: string;
    CAMPO_53_3?: string;
    CAMPO_53_4?: string;
    CAMPO_53_5?: string;
    CAMPO_53_6?: string;
    CAMPO_53_7?: string;
    CAMPO_53_8?: string;
    CAMPO_53_9?: string;
    CAMPO_54?: string;
    CAMPO_55?: string;
    CAMPO_56?: string;
    CAMPO_57?: string;
    CAMPO_58?: string;
    CAMPO_59?: string;
    CAMPO_60?: string;
    CAMPO_61?: string;
    CAMPO_62?: string;
    CAMPO_63?: string;
    CAMPO_64?: string;
    CAMPO_65?: string;
    CAMPO_66?: string;
    CAMPO_66_1?: string;
    CAMPO_66_2?: string;
    CAMPO_66_3?: string;
    CAMPO_66_4?: string;
    CAMPO_66_5?: string;
    CAMPO_66_6?: string;
    CAMPO_66_7?: string;
    CAMPO_66_8?: string;
    CAMPO_66_9?: string;
    CAMPO_67?: string;
    CAMPO_68?: string;
    CAMPO_69?: string;
    CAMPO_70?: string;
    CAMPO_71?: string;
    CAMPO_72?: string;
    CAMPO_73?: string;
    CAMPO_74?: string;
    CAMPO_75?: string;
    CAMPO_76?: string;
    CAMPO_77?: string;
    CAMPO_78?: string;
    CAMPO_79?: string;
    CAMPO_80?: string;
    CAMPO_81?: string;
    CAMPO_82?: string;
    CAMPO_83?: string;
    CAMPO_84?: string;
    CAMPO_85?: string;
    CAMPO_86?: string;
    CAMPO_87?: string;
    CAMPO_88?: string;
    CAMPO_89?: string;
    CAMPO_90?: string;
    CAMPO_91?: string;
    CAMPO_92?: string;
    CAMPO_93?: string;
    CAMPO_94?: string;
    CAMPO_95?: string;
    CAMPO_96?: string;
    CAMPO_97?: string;
    CAMPO_98?: string;
    CAMPO_99?: string;
    CAMPO_100?: string;
    CAMPO_101?: string;
    CAMPO_102?: string;
    CAMPO_103?: string;
    CAMPO_104?: string;
    CAMPO_105?: string;
    CAMPO_106?: string;
    CAMPO_107?: string;
    CAMPO_108?: string;
    CAMPO_109?: string;
    CAMPO_110?: string;
    CAMPO_111?: string;
    CAMPO_112?: string;
    CAMPO_113?: string;
    CAMPO_114?: string;
    CAMPO_114_1?: string;
    CAMPO_114_2?: string;
    CAMPO_114_3?: string;
    CAMPO_114_4?: string;
    CAMPO_114_5?: string;
    CAMPO_114_6?: string;
    CAMPO_115?: string;
    CAMPO_116?: string;
    CAMPO_117?: string;
    CAMPO_118?: string;
    CAMPO_119?: string;
    CAMPO_120?: string;
    CAMPO_121?: string;
    CAMPO_122?: string;
    CAMPO_123?: string;
    CAMPO_124?: string;
    CAMPO_125?: string;
    CAMPO_126?: string;
    CAMPO_127?: string;
    CAMPO_128?: string;
    CAMPO_129?: string;
    CAMPO_130?: string;
    CAMPO_131?: string;
    CAMPO_132?: string;
    CAMPO_133?: string;
    CAMPO_134?: string;

}